requirements.txt
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.4.0
h5py>=3.0.0
pyyaml>=5.4.0
click>=8.0.0
tqdm>=4.60.0